Move boilerplate code out of lengthy module
